Day 17! Aka fluid dynamics for spheres in a vacuum!
Super fun problem, but spent WAYYY too long forgetting to count the columns for
x_min-1 and x_max+1 because I thought the rules were to not count them (similar
the rules for y boundaries). Even though I had purposefully left an extra column
on either side of my grid specifically to account for the possible spillage on
either side! >.<

Also, hilarious that recursion worked here by extending the stack limit to 2500!
I was really worried that this was going to be a poor design decision that would
cost me later. I realized that I could have done it iteratively with stacks,
but recursion was easier to think about and I got too lazy to change my
implementation LOL. Hope that doesn't come back to bite me during part 2!